- Removing Google Map's LocationSource from LocationService.
If Google Map tries to set LocationService as a LocationSource from
LocationActivity, it can happen before LocationService is bound, so
LocationService is null at the time.

Recommended way to do it is: set the fragment/activity that contains
GoogleMap as its LocationSource, and register the fragment/activity
as a common LocationChangedListener through LocationActivity. When
LocationService broadcasts a new location, notify the Google Map
listener from the activity/fragment that contains it.